requirements.txt
python-dotenv
osu.py
